test(agent): mirror context interruption coverage
